# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ipoh Drum Laravel API",
        "description": "This is a swagger-generated API documentation for the project Ipoh Drum. (Only supports OpenAPI Annotations for now.)",
        "contact": {
            "email": "henry_lcz97@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "The URL that IpohDrum Laravel API is running on."
        }
    ],
    "paths": {
...
        "/api/group": {
            "get": {
                "tags": [
                    "GroupControllerService"
                ],
                "summary": "Get list of groups",
                "description": "Returns list of groups",
                "operationId": "getGroups",
                "parameters": [
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "number of pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of groups"
                    },
                    "default": {
                        "description": "Unable to retrieve list of groups"
                    }
                }
            },
            "post": {
                "tags": [
                    "GroupControllerService"
                ],
                "summary": "Creates a group.",
                "operationId": "createGroup",
                "parameters": [
                    {
                        "name": "companyid",
                        "in": "query",
                        "description": "Group belongs to which company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Group Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Group Description",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group has been created successfully."
                    },
                    "default": {
                        "description": "Unable to create the group."
                    }
                }
            }
        },
        "/api/filter/group": {
            "get": {
                "tags": [
                    "GroupControllerService"
                ],
                "summary": "Filter list of groups",
                "description": "Returns list of filtered groups",
                "operationId": "filterGroups",
                "parameters": [
                    {
                        "name": "pageNumber",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "number of pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fromdate",
                        "in": "query",
                        "description": "From Date for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "todate",
                        "in": "query",
                        "description": "To string for filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "status for filter",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of filtered groups"
                    },
                    "default": {
                        "description": "Unable to retrieve list of groups"
                    }
                }
            }
        },
        "/api/group/{uid}": {
            "get": {
                "tags": [
                    "GroupControllerService"
                ],
                "summary": "Retrieves group by Uid.",
                "operationId": "getGroupByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Group_ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group has been retrieved successfully."
                    },
                    "default": {
                        "description": "Unable to retrieve the group."
                    }
                }
            },
            "put": {
                "tags": [
                    "GroupControllerService"
                ],
                "summary": "Update group by Uid.",
                "operationId": "updateGroupByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Group_ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "companyid",
                        "in": "query",
                        "description": "Group belongs to which company",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Groupname",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "Group Description",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group has been updated successfully."
                    },
                    "default": {
                        "description": "Unable to update the group."
                    }
                }
            },
            "delete": {
                "tags": [
                    "GroupControllerService"
                ],
                "summary": "Set group's 'status' to 0.",
                "operationId": "deleteGroupByUid",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "Group ID, NOT 'ID'.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group has been 'deleted' successfully."
                    },
                    "default": {
                        "description": "Unable to 'delete' the group."
                    }
                }
            }
        },
...
    },
...
}